refactor(shared): split diagnostic issue types into issue-diagnostics.ts (FUL-700)
